refactor: abstract getAppInfo
